Try to add reth-primitives

+++ methods/guest/src/main.rs
@@ -4,7 +4,7 @@ fn main() {
     // TODO: Implement your guest code here
 
     // read the input
-    let input: u32 = env::read();
+    let input: u64 = env::read();
 
     // TODO: do something with the input
